Reorganize packages into abi and ui
