+++ app-android/app/proguard-rules.pro
@@ -22,29 +22,29 @@
 
 # JNI entry points are exported with static Java_* symbols, so their Kotlin
 # owner and method names must remain stable after R8.
--keep class com.algoritmico.passepartout.helpers.NativeLibraryWrapper {
+-keep class com.algoritmico.passepartout.abi.NativeLibraryWrapper {
     native <methods>;
 }
 
 # The native core calls these handlers by method name through GetMethodID.
 # Keep both the fun-interface methods and their generated lambda/object
 # implementors so R8 cannot rename the callback entry points.
--keep interface com.algoritmico.passepartout.helpers.ABIEventHandler {
+-keep interface com.algoritmico.passepartout.abi.helpers.ABIEventHandler {
     public void onEvent(java.lang.String);
 }
--keep class * implements com.algoritmico.passepartout.helpers.ABIEventHandler {
+-keep class * implements com.algoritmico.passepartout.abi.helpers.ABIEventHandler {
     public void onEvent(java.lang.String);
 }
--keep interface com.algoritmico.passepartout.helpers.ABIConnectionStatusHandler {
+-keep interface com.algoritmico.passepartout.abi.helpers.ABIConnectionStatusHandler {
     public void onStatus(java.lang.String);
 }
--keep class * implements com.algoritmico.passepartout.helpers.ABIConnectionStatusHandler {
+-keep class * implements com.algoritmico.passepartout.abi.helpers.ABIConnectionStatusHandler {
     public void onStatus(java.lang.String);
 }
--keep interface com.algoritmico.passepartout.helpers.ABICompletionCallback {
+-keep interface com.algoritmico.passepartout.abi.helpers.ABICompletionCallback {
     public void onComplete(int, java.lang.String);
 }
--keep class * implements com.algoritmico.passepartout.helpers.ABICompletionCallback {
+-keep class * implements com.algoritmico.passepartout.abi.helpers.ABICompletionCallback {
     public void onComplete(int, java.lang.String);
 }
