+++ app-android/app/src/main/kotlin/com/algoritmico/passepartout/PassepartoutVPNService.kt
@@ -12,9 +12,8 @@ import androidx.core.app.NotificationChannelCompat
 import androidx.core.app.NotificationCompat
 import androidx.core.app.NotificationManagerCompat
 import com.algoritmico.passepartout.abi.models.OnConnectionStatus
-import com.algoritmico.passepartout.helpers.ABIConnectionStatusDispatcher
-import com.algoritmico.passepartout.helpers.NativeLibraryWrapper
-import com.algoritmico.passepartout.helpers.globalJsonCoder
+import com.algoritmico.passepartout.abi.helpers.ABIConnectionStatusDispatcher
+import com.algoritmico.passepartout.abi.NativeLibraryWrapper
 import io.partout.abi.ConnectionStatus
 import io.partout.abi.TaggedProfile
 import io.partout.jni.AndroidTunnelController
