+++ app-android/app/src/main/kotlin/com/algoritmico/passepartout/abi/helpers/ABIConnectionStatusHandler.kt
@@ -1,4 +1,4 @@
-package com.algoritmico.passepartout.helpers
+package com.algoritmico.passepartout.abi.helpers
 
 fun interface ABIConnectionStatusHandler {
     fun onStatus(onStatusJSON: String)
